Make PollenTest pass with the new handlers architecture
Now PollenTest implements a full round of messages between two vault
users with the new handlers architecture (not doing yet the full
multi-party poll though).

1. Alice sends number poll to Bob (with new RemoteUser)

2. Alice's RemoteUser sends a REQUEST message (now messages have types)
to Bob. This message has a token (useful for the later response). As
Pollen right now is synchronous, Alice waits here, locking herself with
her own SendNumberResponseHandler.

3. Bob's Vaultage receives the message and processes it with
CustomPollenRequestMessageHandler.

4. Bob's CustomPollenRequestMessageHandler calls User.sendNumberPoll()
method, which in this moment just asks for a value and then returns it
(these vault methods need some work in the templates).

5. Bob's CustomPollenRequestMessageHandler prepares a RESPONSE message
with the result of the function called in (4), and the token of the
request message (from 2), and sents it back.

6. Alice's Vaultage receives the message and processes it with
CustomPollenResponseMessageHandler.

7. Alice's CustomPollenResponseMessageHandler then finds the appropriate
operation response handler to call, which in this case is
SendNumberPollResponseHandler. The method of this handler is executed,
which processes the poll result and notifies any threads locked in the
handler object (therefore awakening Alice).

8. Alice retrieves the results of Bob's message and shows the value.