Generate two separate methods for requests and responses
For each vault operation, the following is generated:

- A void method in the @vault class. This method is invoked when a
request message is received.

- A void method in the RemoteVault class. This method can be used to
answer to a previous request message.

Both methods contain an extra "String token" parameter. This token
should be shared between each request/response pair of messages, so that
vaults know what messages they are getting responded (useful when
multiple requests to the same remote vault can happen simultaneously).
To properly work with this token, a vault receiving a request should
store the token and use it when sending back the associated response.